HHVM compat (Travis)
HHVM gives error on !is_callable when the array is converted to string
"Array".

To document PHP default behavior the is_callable php-test has been created
so that the expected outcome for the callable name parameter is known.

Additionally gettype behavior is tested and documented, as those type
names are used in the check for the error-case.

Additionally a little difference in an error message has been streamlined
to not produce a failing test on HHVM.